fixing selection sort files

+++ 2-selection_sort.c
@@ -24,7 +24,7 @@ void swap(int *x, int *y)
 */
 void selection_sort(int *array, size_t size)
 {
-	int i, j, min_index;
+	size_t i, j, min_index;
 	/*after decleration of varaiables*/
 	if (!array || !size)
 		return;
